List: adding hooks to control tasks_quantity; ListService: refactoring (+ connecting hooks) + adding new method

+++ src/modules/lists/list.service.ts
@@ -1,37 +1,50 @@
 import {
-  ForbiddenException,
   HttpException,
   HttpStatus,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
+import { validate } from 'uuid';
 
 import { List } from './list.model';
+import { User } from '../users/user.model';
+
 import { CreateListDto } from './dto/create.list.dto';
 import { UpdateListDto } from './dto/update.list.dto';
-import { User } from '../users/user.model';
 
 @Injectable()
 export class ListService {
   constructor(@InjectModel(List) private listRepository: typeof List) {}
 
   async createList(user: User, dto: CreateListDto): Promise<List> {
-    const list = await this.listRepository.create(dto);
-    if (!list)
-      throw new HttpException('List is not created!', HttpStatus.BAD_REQUEST);
-    const listEntity = await this.listRepository.findOne({
-      where: { user_id: null },
-    });
-    listEntity.user_id = user.id;
-    await listEntity.save();
-    return await this.listRepository.findOne({
-      where: { id: listEntity.id },
-    });
+    const list = await this.listRepository.create(dto, { returning: true });
+    list.user_id = user.id;
+    const updatedList = await list.save();
+    return updatedList;
   }
 
   async getUserListByID(id: string): Promise<List> {
+    if (!validate(id))
+      throw new HttpException(
+        'The id of the user is invalid!',
+        HttpStatus.BAD_REQUEST
+      );
     const list = await this.listRepository.findByPk(id, {
+      include: { all: true },
+    });
+    if (!list) throw new NotFoundException('List not found!');
+    return list;
+  }
+
+  async getUserListByUserId(user: User, id: string): Promise<List> {
+    if (!validate(id))
+      throw new HttpException(
+        'The id of the user is invalid!',
+        HttpStatus.BAD_REQUEST
+      );
+    const list = await this.listRepository.findOne({
+      where: { user_id: user.id, id: id },
       include: { all: true },
     });
     if (!list) throw new NotFoundException('List not found!');
@@ -51,26 +64,34 @@ export class ListService {
     id: string,
     dto: UpdateListDto
   ): Promise<List> {
-    const isListExists = await this.listRepository.findOne({
+    if (!validate(id))
+      throw new HttpException(
+        'The id of the user is invalid!',
+        HttpStatus.BAD_REQUEST
+      );
+    const list = await this.listRepository.findOne({
       where: { user_id: user.id, id: id },
     });
-    if (!isListExists) throw new NotFoundException('List not found!');
-    this.listRepository.update({ title: dto.title }, { where: { id: id } });
-    return await this.listRepository.findOne({
-      where: { id: id },
-      include: { all: true },
-    });
+    if (!list) throw new NotFoundException('List not found!');
+    await list.update({ title: dto.title }, { silent: false });
+    await list.reload();
+    return list;
   }
 
   async deleteUserListById(user: User, id: string): Promise<void> {
+    if (!validate(id))
+      throw new HttpException(
+        'The id of the user is invalid!',
+        HttpStatus.BAD_REQUEST
+      );
     const listOwnership = await this.listRepository.findOne({
       where: { user_id: user.id, id: id },
     });
     if (!listOwnership)
-      throw new ForbiddenException('You do not have such a list');
-    const list = await this.listRepository.destroy({
+      throw new NotFoundException('There is no such list in your account!');
+    await this.listRepository.destroy({
       where: { id: id },
+      individualHooks: true,
     });
-    if (list === 0) throw new NotFoundException('List not found');
   }
 }
